pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "public-docs-llm-kit"
version = "0.1.0"
description = "Grounded public documentation pipeline MVP with local retrieval and citations."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "fastapi>=0.115.0",
  "httpx>=0.28.0",
  "markdownify>=0.13.1",
  "numpy>=2.1.0",
  "openai>=2.29.0",
  "playwright>=1.52.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.6.0",
  "pyyaml>=6.0.2",
  "uvicorn>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[project.scripts]
public-docs-llm = "public_docs_llm.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
